# api/openapi.yaml
openapi: 3.0.3
info:
  title: Tenara Platform API
  version: 0.1.0
  description: >
    Single API surface shared by Web Console, CLI and MCP.
    Contract only exposes platform-level concepts; infrastructure primitives
    (namespaces, pods, services, ingress, network policies) are never visible.
servers:
  - url: http://localhost:8080
    description: local control plane
...
security:
  - bearerAuth: []
paths:
...
  /v1/apps/{appId}:
    get:
      operationId: GetV1AppsByAppId
      tags: [apps]
      summary: get one application
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: app
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: PatchV1AppsByAppId
      tags: [apps]
      summary: update mutable fields
      parameters:
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppUpdateInput" }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/App" }
    delete:
      operationId: DeleteV1AppsByAppId
      tags: [apps]
      summary: soft delete (async pipeline, restorable during retention)
      parameters:
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: deletion pipeline started
        "404":
          $ref: "#/components/responses/Problem"
...
  /v1/analyze:
    post:
      operationId: PostV1Analyze
      tags: [analyze]
      summary: scan a repository and produce an AppSpec proposal
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnalyzeInput" }
      responses:
        "200":
          description: proposed AppSpec with confidence report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnalyzeResult" }
        "422":
          $ref: "#/components/responses/UnsupportedStackProblem"
...
  /v1/apps/{appId}/deployments:
    post:
      operationId: PostV1AppsDeployments
      tags: [deployments]
      summary: approve a plan and start a deployment (plan_id is the approval)
      parameters:
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id]
              properties:
                plan_id: { type: string }
      responses:
        "202":
          description: deployment started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
        "410":
          $ref: "#/components/responses/PlanExpiredProblem"
        "402":
          $ref: "#/components/responses/QuotaProblem"
  /v1/apps/{appId}/deployments/{deploymentId}:
    get:
      operationId: GetV1AppsDeployments
      tags: [deployments]
      summary: deployment status including diagnostics reference when degraded
      parameters:
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/DeploymentId"
      responses:
        "200":
          description: status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
        "404":
          $ref: "#/components/responses/Problem"
...
  /v1/apps/{appId}/databases:
    post:
      operationId: PostV1AppsDatabases
      tags: [databases]
      summary: request an isolated database binding
      parameters:
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DatabaseRequest" }
      responses:
        "201":
          description: binding pending provisioning
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DatabaseBinding" }
        "402":
          $ref: "#/components/responses/QuotaProblem"
  /v1/apps/{appId}/diagnostics:
    get:
      operationId: GetV1AppsDiagnostics
      tags: [deployments]
      summary: structured failure diagnostics for the latest failed verification
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: diagnostics bundle
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Diagnostics" }
...
  /v1/admin/users:
    get:
      operationId: GetV1AdminUsers
      tags: [admin]
      summary: list users (platform admin capability)
      responses:
        "200":
          description: users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "403":
          $ref: "#/components/responses/Problem"
  /v1/admin/users/{userId}/suspend:
    post:
      operationId: PostV1AdminUsersSuspend
      tags: [admin]
      summary: suspend or restore a user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspended]
              properties:
                suspended: { type: boolean }
      responses:
        "204":
          description: applied
        "403":
          $ref: "#/components/responses/Problem"
  /v1/admin/apps:
    get:
      operationId: GetV1AdminApps
      tags: [admin]
      summary: list all applications across workspaces
      responses:
        "200":
          description: apps
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/App" }
        "403":
          $ref: "#/components/responses/Problem"
  /v1/admin/quota:
    put:
      operationId: PutV1AdminQuota
      tags: [admin]
      summary: override quota tier for an organization
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org_id, tier]
              properties:
                org_id: { type: string }
                tier: { type: string, enum: [free, pro] }
      responses:
        "204":
          description: updated
        "403":
          $ref: "#/components/responses/Problem"
  /v1/admin/security-events:
    get:
      operationId: GetV1AdminSecurityEvents
      tags: [admin]
      summary: stream security events
      responses:
        "200":
          description: events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SecurityEvent" }
        "403":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: session JWT or workspace API token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 255 }
      description: replay-safe key for mutation requests
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string }
    DeploymentId:
      name: deploymentId
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: RFC7807 problem
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    QuotaProblem:
      description: quota exceeded (payment required semantics)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedStackProblem:
      description: analyzer could not map the stack
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    PlanExpiredProblem:
      description: plan approval window elapsed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        error_code:
          type: string
          enum:
            - QUOTA_EXCEEDED
            - UNSUPPORTED_STACK
            - ZERO_SERVICES
            - PLAN_EXPIRED
            - DOMAIN_NOT_VERIFIED
            - IDEMPOTENCY_CONFLICT
            - VALIDATION_FAILED
            - FORBIDDEN_CAPABILITY
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        email_verified: { type: boolean }
        github_bound: { type: boolean }
...
    App:
      type: object
      required: [id, name, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        deleted_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
    AppUpdateInput:
      type: object
      properties:
        name: { type: string }
...
    AppSpec:
      type: object
      required: [version, services]
      properties:
        version:
          type: string
          enum: [v1]
        services:
          type: object
          additionalProperties:
            type: object
            required: [type, runtime, path]
            properties:
              type: { type: string, enum: [frontend, backend] }
              runtime: { type: string, enum: [node, python, go] }
              path: { type: string }
              port: { type: integer }
        database:
          type: object
          properties:
            mongodb: { type: boolean }
        routing:
          type: object
          additionalProperties:
            type: object
            required: [service]
            properties:
              service: { type: string }
    AnalyzeInput:
      type: object
      required: [repo_url]
      properties:
        repo_url: { type: string }
        ref: { type: string, default: main }
        github_token_ref: { type: string }
    ConfidenceEntry:
      type: object
      required: [field, decision, evidence]
      properties:
        field: { type: string }
        decision: { type: string }
        evidence:
          type: array
          items: { type: string }
    AnalyzeResult:
      type: object
      required: [app_spec, confidence]
      properties:
        app_spec: { $ref: "#/components/schemas/AppSpec" }
        confidence:
          type: array
          items: { $ref: "#/components/schemas/ConfidenceEntry" }
...
    DeploymentState:
      type: string
      enum:
        - CREATED
        - AWAITING_APPROVAL
        - PLANNED
        - QUEUED
        - BUILDING
        - DEPLOYING
        - VERIFYING
        - RUNNING
        - DEGRADED
        - FAILED
        - STOPPED
        - ROLLING_BACK
        - DELETING
        - DELETED
        - RESTORED
    Deployment:
      type: object
      required: [id, app_id, state]
      properties:
        id: { type: string }
        app_id: { type: string }
        state: { $ref: "#/components/schemas/DeploymentState" }
        git_sha: { type: string }
        image_digest: { type: string }
        revision: { type: integer }
        diagnostics_ref: { type: string }
        created_at: { type: string, format: date-time }
...
    DatabaseRequest:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [mongodb] }
        isolation: { type: string, enum: [shared, dedicated], default: shared }
    DatabaseBinding:
      type: object
      required: [id, app_id, type, state]
      properties:
        id: { type: string }
        app_id: { type: string }
        type: { type: string }
        isolation: { type: string }
        state: { type: string, enum: [pending, ready, failed] }
...
    SecurityEvent:
      type: object
      required: [kind, source_ip, timestamp]
      properties:
        kind: { type: string }
        source_ip: { type: string }
        detail: { type: string }
        timestamp: { type: string, format: date-time }
    Diagnostics:
      type: object
      required: [classified]
      properties:
        classified: { type: string }
        suggestions:
          type: array
          items: { type: string }
        build_log_tail: { type: string }
        pod_events: { type: string }